requirements.txt
gitpython
ipython 
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
requests>=2.23.0
torch==1.12.1
torchvision==0.13.1
tqdm>=4.64.0
protobuf<=3.20.1 
pandas>=1.1.4

